Overhaul Mark 1, first step to moving the cpns forward
There were some unused tables/models and default and mobile numbers were
seperate from ppc numbers. To get to where we are heading with the phone
number service the seperation would make things more complicated. This
combines those two tables, adds some test coverage and removes other
stuff.

There will need to be a second deploy that removes the PpcNumber Model,
because it is used in one of the migrations.